requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
tensorflow
keras
mlflow
fastapi
ipykernel
boto3
pytest
pyspark
apache-airflow
python-dotenv
pyyaml
xgboost